Require input data to sst_anomaly to include ref.years
The sst_anomaly() function was re-written to require that the input data
to the function contain all the years to calculate the long-term mean
(ref.years). Having this requirement allows the function to be more
flexible in terms of the spatial and temporal subsets of data that be
used to calculate anomalies.